requirements.txt
asgiref==3.8.1
Django==5.0.6
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.8.0
python-magic==0.4.27
python-magic-bin==0.4.14
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1